+++ examples/Infrastructure/Model/ClassStudentModel.php
@@ -16,9 +16,9 @@ class ClassStudentModel extends AbstractModel
         return "classes_students";
     }
 
-    public function getPrimaryKey(): string
+    public function getPrimaryKeys(): array
     {
-        return $this->id;
+        return [$this->id];
     }
 
     protected function getRelationships(): array