+++ src/Execution/Persister.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 namespace WoohooLabs\Worm\Execution;
 
 use WoohooLabs\Larva\Connection\ConnectionInterface;
-use WoohooLabs\Larva\Query\Condition\ConditionBuilder;
 use WoohooLabs\Larva\Query\Delete\DeleteQueryBuilder;
 use WoohooLabs\Larva\Query\Insert\InsertQueryBuilder;
 use WoohooLabs\Larva\Query\Update\UpdateQueryBuilder;
@@ -35,10 +34,10 @@ class Persister
     public function save(ModelInterface $model, array $record, $entity)
     {
         $type = $model->getTable();
-        $id = $model->getId($record);
+        $hash = $model->getHash($record);
 
-        if ($this->identityMap->getState($type, $id) === IdentityMap::STATE_MANAGED) {
-            $this->update($model, $record, $entity);
+        if ($this->identityMap->getState($type, $hash) === IdentityMap::STATE_MANAGED) {
+            $this->update($model, $record, $hash, $entity);
         } else {
             $this->insert($model, $record, $entity);
         }
@@ -65,21 +64,23 @@ class Persister
         $relatedEntity
     ) {
         $type = $model->getTable();
-        $id = $model->getId($record);
+        $hash = $model->getHash($record);
 
         $relatedModel = $model->getRelationship($relationship)->getModel();
         $relatedType = $relatedModel->getTable();
+        $relatedHash = $relatedModel->getHash($relatedRecord);
         $relatedId = $relatedModel->getId($relatedRecord);
 
-        if ($this->identityMap->hasId($relatedType, $relatedId)) {
-            if ($this->identityMap->hasRelatedId($type, $id, $relationship, $relatedId)) {
-                $this->update($relatedModel, $relatedRecord, $relatedEntity);
+        if ($this->identityMap->hasIdentity($relatedType, $relatedHash)) {
+            if ($this->identityMap->hasRelatedIdentity($type, $hash, $relationship, $relatedHash)) {
+                $this->update($relatedModel, $relatedRecord, $relatedHash, $relatedEntity);
             } else {
-                $this->identityMap->removeRelatedId($type, $id, $relationship, $relatedId);
+                $this->identityMap->removeRelatedIdentity($type, $hash, $relationship, $relatedHash);
+                $relatedId = $relatedModel->getId($relatedRecord);
                 $this->delete($relatedModel, $relatedId);
             }
         } else {
-            $this->identityMap->addRelatedId($type, $id, $relationship, $relatedType, $relatedId);
+            $this->identityMap->addRelatedIdentity($type, $hash, $relationship, $relatedType, $relatedHash, $relatedId);
             $this->insert($relatedModel, $relatedRecord, $relatedEntity);
         }
     }
@@ -96,19 +97,18 @@ class Persister
         $relatedObjects
     ) {
         $relatedModel = $model->getRelationship($relationship)->getModel();
-        $relatedIds = [];
 
+        $relatedIds = [];
         foreach ($relatedRecords as $key => $relatedRecord) {
-            $relatedId = $relatedModel->getId($relatedRecord);
             $relatedObject = $relatedObjects[$key] ?? null;
 
             $this->saveRelatedEntity($model, $relationship, $record, $relatedRecord, $relatedObject);
-            $relatedIds[] = $relatedId;
+            $relatedIds[] = $relatedModel->getId($relatedRecord);
         }
 
         $type = $model->getTable();
-        $id = $model->getId($record);
-        $deletedIds = array_diff($this->identityMap->getRelatedIds($type, $id, $relationship), $relatedIds);
+        $hash = $model->getHash($record);
+        $deletedIds = array_diff($this->identityMap->getRelatedIds($type, $hash, $relationship), $relatedIds);
         foreach ($deletedIds as $deletedId) {
             $this->delete($relatedModel, $deletedId);
         }
@@ -133,34 +133,31 @@ class Persister
         $query->execute($this->connection);
 
         $type = $model->getTable();
-        $id = $model->getId($record);
-        $this->identityMap->addId($type, $id, $entity);
+        $hash = $model->getHash($record);
+        $this->identityMap->addIdentity($type, $hash, $entity);
     }
 
     /**
      * @param object|null $entity
      * @return void
      */
-    private function update(ModelInterface $model, array $record, $entity)
+    private function update(ModelInterface $model, array $record, string $hash, $entity)
     {
         $type = $model->getTable();
         $id = $model->getId($record);
 
-        if ($this->identityMap->getState($type, $id) !== IdentityMap::STATE_MANAGED) {
+        if ($this->identityMap->getState($type, $hash) !== IdentityMap::STATE_MANAGED) {
             return;
         }
 
         $query = UpdateQueryBuilder::create()
             ->table($model->getTable())
             ->setValues($record)
-            ->where(
-                ConditionBuilder::create()
-                    ->columnToValue($model->getPrimaryKey(), "=", $model->getId($record))
-            );
+            ->where($model->createConditionBuilder($id));
 
         $query->execute($this->connection);
 
-        $this->identityMap->setObject($type, $id, $entity);
+        $this->identityMap->setObject($type, $hash, $entity);
     }
 
     /**
@@ -170,22 +167,20 @@ class Persister
     private function doDelete(ModelInterface $model, $id)
     {
         $type = $model->getTable();
+        $hash = $model->getHashFromId($id);
 
-        if ($this->identityMap->getState($type, $id) !== IdentityMap::STATE_MANAGED) {
+        if ($this->identityMap->getState($type, $hash) !== IdentityMap::STATE_MANAGED) {
             return;
         }
 
         $query = DeleteQueryBuilder::create()
             ->from($model->getTable())
-            ->where(
-                ConditionBuilder::create()
-                    ->columnToValue($model->getPrimaryKey(), "=", $id)
-            );
+            ->where($model->createConditionBuilder($id));
 
         $query->execute($this->connection);
 
-        $this->identityMap->setState($type, $id, IdentityMap::STATE_DELETED);
-        $this->identityMap->setObject($type, $id, null);
+        $this->identityMap->setState($type, $hash, IdentityMap::STATE_DELETED);
+        $this->identityMap->setObject($type, $hash, null);
         $model->cascadeDelete($this, $id);
     }
 }