+++ src/Model/ModelInterface.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace WoohooLabs\Worm\Model;
 
 use DomainException;
+use WoohooLabs\Larva\Query\Condition\ConditionBuilderInterface;
 use WoohooLabs\Worm\Execution\IdentityMap;
 use WoohooLabs\Worm\Execution\Persister;
 use WoohooLabs\Worm\Model\Relationship\RelationshipInterface;
@@ -12,7 +13,10 @@ interface ModelInterface
 {
     public function getTable(): string;
 
-    public function getPrimaryKey(): string;
+    /**
+     * @return string[]
+     */
+    public function getPrimaryKeys(): array;
 
     /**
      * @return string[]
@@ -25,9 +29,21 @@ interface ModelInterface
     public function getRelationship(string $name): RelationshipInterface;
 
     /**
-     * @return mixed
+     * @return mixed|mixed[]|null
      */
     public function getId(array $record);
+
+    public function getHash(array $record): string;
+
+    /**
+     * @param mixed $id
+     */
+    public function getHashFromId($id): string;
+
+    /**
+     * @param mixed $id
+     */
+    public function createConditionBuilder($id): ConditionBuilderInterface;
 
     /**
      * @return void