+++ src/Query/UpdateQueryBuilder.php
@@ -70,10 +70,8 @@ class UpdateQueryBuilder
      */
     public function executeById($id): bool
     {
-        $this->addWhereGroup(
-            ConditionBuilder::create()
-                ->columnToValue($this->model->getPrimaryKey(), "=", $id)
-        );
+        $this->queryBuilder
+            ->addWhereGroup($this->model->createConditionBuilder($id));
 
         return $this->queryExecutor->update($this->queryBuilder);
     }